Split ConsumerConfig into two classes. - ConsumerConfig and ConsumerGroupConfig
